+++ app/controllers/flow/outgoings_handler.rb
@@ -10,12 +10,6 @@ module Flow
       def form(params, _session_data)
         OutgoingsForm.new(params.require(:outgoings_form).permit(*OutgoingsForm::OUTGOING_ATTRIBUTES, outgoings: []))
       end
-
-      # def save_data(cfe_connection, estimate_id, outgoings_form, session_data)
-      #   income_form = MonthlyIncomeHandler.model(session_data)
-      #
-      #   cfe_connection.create_regular_payments(estimate_id, income_form, outgoings_form)
-      # end
     end
   end
 end
