+++ app/controllers/flow/property_entry_handler.rb
@@ -12,15 +12,6 @@ module Flow
           model.property_owned = session_data["property_owned"]
         end
       end
-
-      # def save_data(cfe_connection, estimate_id, model, session_data)
-      #   main_home = {
-      #     value: model.house_value,
-      #     outstanding_mortgage: (model.mortgage.presence if session_data["property_owned"] == "with_mortgage") || 0,
-      #     percentage_owned: model.percentage_owned,
-      #   }
-      #   cfe_connection.create_properties(estimate_id, main_home, nil)
-      # end
     end
   end
 end
