+++ app/controllers/flow/property_handler.rb
@@ -10,9 +10,6 @@ module Flow
       def form(params, _session_data)
         PropertyForm.new(params.require(:property_form).permit(:property_owned))
       end
-
-      # Will be called when no property owned
-      # def save_data(cfe_connection, estimate_id, estimate, _session_data); end
     end
   end
 end
