+++ app/controllers/flow/vehicle/owned_handler.rb
@@ -9,9 +9,6 @@ module Flow
         def form(params, _session_data)
           VehicleForm.new params.require(:vehicle_form).permit(*VehicleForm::VEHICLE_ATTRIBUTES)
         end
-
-        # called when 'no vehicle' selected - nothing to do
-        # def save_data(cfe_connection, estimate_id, estimate, _session_data); end
       end
     end
   end
