+++ spec/controllers/estimates_controller_spec.rb
@@ -1,0 +1,29 @@
+# require 'rails_helper'
+#
+# RSpec.describe EstimatesController  type: :controller do
+#
+#
+#   # expect(mock_connection).to receive(:create_employment) do |_id, params|
+#   #   payment = params.dig(0, :payments, 1)
+#   #   expect(payment[:gross]).to eq 1_000
+#   #   expect(payment[:date]).to eq 2.weeks.ago.to_date
+#   # end
+#   #
+#   # it "persists my answers to CFE and moves me on to the next question" do
+#   #   expect(mock_connection).to receive(:create_employment) do |id, params|
+#   #     expect(id).to eq estimate_id
+#   #     payment = params.dig(0, :payments, 1)
+#   #     expect(payment[:gross]).to eq 1_000
+#   #     expect(payment[:tax]).to eq(-100)
+#   #     expect(payment[:national_insurance]).to eq(-50)
+#   #     expect(payment[:date]).to eq 1.month.ago.to_date
+#   #   end
+#   #
+#   #   click_on "Save and continue"
+#   #   expect(page).not_to have_content employment_page_header
+#   # end
+# end
+
+
+
+
